### 2022-01-18 #4073 @whedon
[REVIEW]: Elephas: Distributed Deep Learning with Keras & Spark
**Submitting author:** @maxpumperla (<a href="http://orcid.org/0000-0002-7801-4184">Max Pumperla</a>)
**Repository:** https://github.com/danielenricocahall/elephas
**Branch with paper.md** (empty if default branch): mp_joss_paper
**Version:** 3.4.7
