### 2022-08-22 #4073 @burch-cm
Hello, I got the email, and I've been unable to install the software or resolve the dependency issues. While I am still interested in helping with reviews, I don't currently have the technical expertise to troubleshoot the issues with Spark/PySpark/Java and I'm sure this will cause even more delays.

Please remove me as a reviewer at this time. Thank you.
