+++ .github/workflows/ci.yaml
@@ -4,28 +4,26 @@ on: [push]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version: ["3.7", "3.8", "3.9"]
-
     steps:
       - uses: actions/checkout@v3
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install Other dependencies
+      - name: Install Spark dependencies
         run: |
           sudo add-apt-repository -y ppa:openjdk-r/ppa
           sudo apt-get -qq update
           sudo apt-get install -y openjdk-8-jdk --no-install-recommends
           sudo update-java-alternatives -s java-1.8.0-openjdk-amd64
           export JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64
+          export SPARK_LOCAL_IP=0.0.0.0
       - name: Install dependencies
         run: |
-          export SPARK_LOCAL_IP=0.0.0.0
           python -m pip install --upgrade pip
           pip install -e .[tests]
       - name: Test with pytest
