+++ README.md
@@ -4,7 +4,7 @@
 
 ## 
 
-[![Build Status](https://app.travis-ci.com/danielenricocahall/elephas.svg?branch=master)](https://app.travis-ci.com/danielenricocahall/elephas)
+[![Build Status](https://github.com/danielenricocahall/elephas/actions/workflows/ci.yaml/badge.svg)](https://github.com/danielenricocahall/elephas/actions/workflows/ci.yaml/badge.svg)
 [![license](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](https://github.com/danielenricocahall/elephas/blob/master/LICENSE)
 
 Elephas is an extension of [Keras](http://keras.io), which allows you to run distributed deep learning models at 
