### 2022-10-10 ca3b380
fix bug with example
