+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 from setuptools import find_packages
 
 setup(name='elephas',
-      version='3.4.0',
+      version='3.4.1',
       description='Deep learning on Spark with Keras',
       url='http://github.com/danielenricocahall/elephas',
       download_url='https://github.com/danielenricocahall/elephas/tarball/3.4.0',
