### 2022-10-10 d1b4d52
Update README.md
revising docs to reflect the fact that we no longer support hyperparam opt.
