### 2022-10-10 8fb73ec
Update README.md
add distributed inference/evaluation as a feature
