### 2022-10-16 351b024
implement mkdocs GitHub Action for auto-documentation
