+++ docs/templates/models/hyper-param-model.md
@@ -1,4 +1,0 @@
-# Distributed hyperparameter search
-
-
-{{autogenerated}}
