+++ .github/workflows/docs.yaml
@@ -16,6 +16,6 @@ jobs:
         uses: mhausenblas/mkdocs-deploy-gh-pages@master
         # Or use mhausenblas/mkdocs-deploy-gh-pages@nomaterial to build without the mkdocs-material theme
         env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          GITHUB_TOKEN: ${{ secrets.PERSONAL_TOKEN }}
           CONFIG_FILE: docs/mkdocs.yml
           REQUIREMENTS: requirements.txt
