### 2022-10-16 c3947a0
update auto-documentation steps
