+++ .github/workflows/docs.yaml
@@ -9,6 +9,17 @@ jobs:
     name: Deploy docs
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python 3.9
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.9"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install .
+
       - name: Checkout main
         uses: actions/checkout@v2
 
