### 2022-10-16 4b9fe9c
update auto-documentation steps
