+++ .github/workflows/docs.yaml
@@ -12,4 +12,4 @@ jobs:
         with:
           python-version: 3.9
       - run: pip install mkdocs
-      - run: mkdocs gh-deploy --force --clean --verbose
+      - run: mkdocs gh-deploy --config-file docs/mkdocs.yml --force --clean --verbose
