### 2022-10-16 078674e
update MkDocs configuration
