+++ docs/mkdocs.yml
@@ -1,13 +1,13 @@
 site_name: Elephas Documentation
 theme: readthedocs
-docs_dir: sources
+docs_dir: docs
 repo_url: http://github.com/Yoseph-Zuskin/elephas
 site_url: http://Yoseph-Zuskin.github.io/elephas
 site_description: 'Documentation for Elephas, Distributed Deep Learning with Keras & Spark'
 
 dev_addr: '0.0.0.0:8000'
 
-pages:
+nav:
 - Home: index.md
 - Why use Elephas: why-use-elephas.md
 - Getting started:
