### 2022-10-16 8a46444
update MkDocs configuration
