+++ .github/workflows/docs.yaml
@@ -1,4 +1,4 @@
-name: build
+name: MkDocs documentation
 on:
   push:
     branches:
@@ -11,5 +11,9 @@ jobs:
       - uses: actions/setup-python@v2
         with:
           python-version: 3.9
-      - run: pip install mkdocs
-      - run: mkdocs gh-deploy --config-file docs/mkdocs.yml --force --clean --verbose
+      - name: Install MkDocs
+        run: pip install mkdocs
+      - name: Build documentation
+        run: mkdocs build -f docs/mkdocs.yml -d docs/sources -c -v
+      - name: Deploy documentation
+        run: mkdocs gh-deploy -f docs/mkdocs.yml -c -v --force
