### 2022-10-16 1b7fa6f
update MkDocs configuration
