### 2022-10-16 440a38c
show contents of site directory
