### 2022-10-16 c417662
update MkDocs configuration
