### 2022-10-16 e3a11fb
update MkDocs auto-documentation steps
