### 2022-10-16 6ac40d9
update MkDocs configuration
