### 2022-10-16 d0b9a62
update MkDocs auto-documentation steps
