+++ .github/workflows/docs.yaml
@@ -14,10 +14,10 @@ jobs:
       - name: Install MkDocs
         run: pip install mkdocs
       - name: Make site directory
-        run: mkdir docs/sources
+        run: mkdir sources
       - name: Build documentation
         run: |
           mkdocs build -f docs/mkdocs.yml -d sources -c -v
-          ls -al docs/sources
+          ls -al sources
       - name: Deploy documentation
         run: mkdocs gh-deploy -f docs/mkdocs.yml -c -v --force
