### 2022-10-16 4d026b3
update MkDocs configuration
