+++ docs/mkdocs.yml
@@ -1,6 +1,7 @@
 site_name: Elephas Documentation
 theme: readthedocs
-site_dir: sources
+docs_dir: templates
+site_dir: ../sources
 repo_url: http://github.com/Yoseph-Zuskin/elephas
 site_url: http://Yoseph-Zuskin.github.io/elephas
 site_description: 'Documentation for Elephas, Distributed Deep Learning with Keras & Spark'
