+++ .github/workflows/docs.yaml
@@ -13,6 +13,7 @@ jobs:
           python-version: 3.9
       - name: Install dependencies
         run: |
+          python -m pip install --upgrade pip
           pip install .
           pip install mkdocs
       - name: Build documentation
