+++ docs/autogen.py
@@ -4,6 +4,7 @@ import re
 import inspect
 import os
 import shutil
+from pathlib import Path
 
 from elephas import spark_model, ml_model
 from elephas.parameter import client, server
@@ -356,7 +357,7 @@ def read_page_data(page_data, type):
 
 
 if __name__ == '__main__':
-    readme = read_file('../README.md')
+    readme = read_file(Path(__file__).parents[1]/'README.md')
     index = read_file('templates/index.md')
     index = index.replace('{{autogenerated}}', readme[readme.find('##'):])
     with open('sources/index.md', 'w') as f:
