+++ docs/autogen.py
@@ -357,10 +357,11 @@ def read_page_data(page_data, type):
 
 
 if __name__ == '__main__':
-    readme = read_file(Path(__file__).parents[1]/'README.md')
-    index = read_file('templates/index.md')
+    pwd = Path(__file__).parent
+    readme = read_file(pwd.parent/'README.md')
+    index = read_file(pwd/'templates/index.md')
     index = index.replace('{{autogenerated}}', readme[readme.find('##'):])
-    with open('sources/index.md', 'w') as f:
+    with open(pwd/'sources/index.md', 'w') as f:
         f.write(index)
 
     print('Generating Elephas docs')
@@ -412,17 +413,17 @@ if __name__ == '__main__':
         # or create page otherwise
         page_name = page_data['page']
         path = os.path.join('sources', page_name)
-        if os.path.exists(path):
-            template = read_file(path)
+        if os.path.exists(pwd/path):
+            template = read_file(pwd/path)
             assert '{{autogenerated}}' in template, ('Template found for ' + path +
                                                      ' but missing {{autogenerated}}'
                                                      ' tag.')
             mkdown = template.replace('{{autogenerated}}', mkdown)
-            print('...inserting autogenerated content into template:', path)
+            print('...inserting autogenerated content into template:', pwd/path)
         else:
-            print('...creating new page with autogenerated content:', path)
-        subdir = os.path.dirname(path)
-        if not os.path.exists(subdir):
-            os.makedirs(subdir)
-        with open(path, 'w') as f:
+            print('...creating new page with autogenerated content:', pwd/path)
+        subdir = os.path.dirname(pwd/path)
+        if not os.path.exists(pwd/subdir):
+            os.makedirs(pwd/subdir)
+        with open(pwd/path, 'w') as f:
             f.write(mkdown)
