+++ docs/autogen.py
@@ -23,6 +23,7 @@ EXCLUDE = {
     'SocketServer'
 }
 
+PWD = Path(__file__).parent
 
 PAGES = [
     {
@@ -90,7 +91,7 @@ PAGES = [
     },
 ]
 
-ROOT = 'http://maxpumperla.com/elephas'
+ROOT = 'http://danielenricocahall.com/elephas'
 
 
 def get_function_signature(function, method=True):
@@ -155,7 +156,7 @@ def class_to_source_link(cls):
     path = module_name.replace('.', '/')
     path += '.py'
     line = inspect.getsourcelines(cls)[-1]
-    link = ('https://github.com/maxpumperla/'
+    link = ('https://github.com/danielenricocahall/'
             'elephas/blob/master/' + path + '#L' + str(line))
     return '[[source]](' + link + ')'
 
@@ -290,20 +291,21 @@ def process_docstring(docstring):
             '$CODE_BLOCK_%d' % i, code_block)
     return docstring
 
-print('Cleaning up existing sources directory.')
-if os.path.exists('sources'):
-    shutil.rmtree('sources')
+
+if os.path.exists(PWD/'sources'):
+    print('Cleaning up existing sources directory.')
+    shutil.rmtree(PWD/'sources')
 
 print('Populating sources directory with templates.')
-for subdir, dirs, fnames in os.walk('templates'):
+for subdir, dirs, fnames in os.walk(PWD/'templates'):
     for fname in fnames:
-        new_subdir = subdir.replace('templates', 'sources')
-        if not os.path.exists(new_subdir):
-            os.makedirs(new_subdir)
+        new_subdir = subdir.replace(str(PWD/'templates'), str(PWD/'sources'))
+        if not os.path.exists(PWD/new_subdir):
+            os.makedirs(PWD/new_subdir)
         if fname[-3:] == '.md':
-            fpath = os.path.join(subdir, fname)
+            fpath = os.path.join(PWD/subdir, PWD/fname)
             new_fpath = fpath.replace('templates', 'sources')
-            shutil.copy(fpath, new_fpath)
+            shutil.copy(PWD/fpath, PWD/new_fpath)
 
 
 def read_file(path):
@@ -357,11 +359,10 @@ def read_page_data(page_data, type):
 
 
 if __name__ == '__main__':
-    pwd = Path(__file__).parent
-    readme = read_file(pwd.parent/'README.md')
-    index = read_file(pwd/'templates/index.md')
+    readme = read_file(PWD.parent/'README.md')
+    index = read_file(PWD/'templates/index.md')
     index = index.replace('{{autogenerated}}', readme[readme.find('##'):])
-    with open(pwd/'sources/index.md', 'w') as f:
+    with open(PWD/'sources/index.md', 'w') as f:
         f.write(index)
 
     print('Generating Elephas docs')
@@ -412,18 +413,18 @@ if __name__ == '__main__':
         # Either insert content into existing page,
         # or create page otherwise
         page_name = page_data['page']
-        path = os.path.join('sources', page_name)
-        if os.path.exists(pwd/path):
-            template = read_file(pwd/path)
+        path = os.path.join(PWD/'sources', page_name)
+        if os.path.exists(path):
+            template = read_file(path)
             assert '{{autogenerated}}' in template, ('Template found for ' + path +
                                                      ' but missing {{autogenerated}}'
                                                      ' tag.')
             mkdown = template.replace('{{autogenerated}}', mkdown)
-            print('...inserting autogenerated content into template:', pwd/path)
+            print('...inserting autogenerated content into template:', path)
         else:
-            print('...creating new page with autogenerated content:', pwd/path)
[…]
