### 2022-10-17 c733c14
update function signature handling
