+++ docs/autogen.py
@@ -100,29 +100,14 @@ def get_function_signature(function, method=True):
         signature = inspect.signature(function)
     else:
         signature = inspect.signature(wrapped)
-    defaults = signature.defaults
+    parameters = [str(p[1]) for p in signature.parameters.items()]
     if method:
-        args = signature.args[1:]
-    else:
-        args = signature.args
-    if defaults:
-        kwargs = zip(args[-len(defaults):], defaults)
-        args = args[:-len(defaults)]
-    else:
-        kwargs = []
+        parameters = parameters[1:]
     st = '%s.%s(' % (clean_module_name(function.__module__), function.__name__)
 
-    for a in args:
-        st += str(a) + ', '
-    for a, v in kwargs:
-        if isinstance(v, str):
-            v = '\'' + v + '\''
-        st += str(a) + '=' + str(v) + ', '
-    if kwargs or args:
-        signature = st[:-2] + ')'
-    else:
-        signature = st + ')'
-    return signature
+    for p in parameters:
+        st += str(p) + ', '
+    return st + ')'
 
 
 def get_class_signature(cls):
