+++ .github/workflows/ci.yaml
@@ -1,6 +1,11 @@
 name: Python package
 
-on: [push]
+on:
+  push:
+    branches:
+      - "**"
+      - "!gh-pages"
+      - "!mkdocs"
 
 jobs:
   build:
