### 2022-10-17 8a2420b
update function signature handling
