### 2022-10-17 eee32a3
update GitHub pages deployment steps
