### 2022-10-17 d17b9ab
update GitHub pages deployment steps
