+++ .github/workflows/ci.yaml
@@ -5,7 +5,6 @@ on:
     branches:
       - "**"
       - "!gh-pages"
-      - "!mkdocs"
 
 jobs:
   build:
