### 2022-10-17 8a12197
remove `ls -al` command from documentation building
