### 2022-10-18 #4073 @maxpumperla
@diehlpk @arfon I could go through the list of reviewers yet again, but I don't think I have any relation to anyone on it. Isn't the appointment of reviewers something you (as part of the journal) are better equipped to handle?
