### 2022-10-24 #4073 @sepandhaghighi
> @sepandhaghighi would you mind reviewing this paper?
💯
Why not?
I'm available for the reviewing
