### 2022-10-26 #4073 @diehlpk
> @sepandhaghighi moving forward this project will be maintained by Daniel Cahill. He decided to create this own fork. The important bit is that a) the info on the repo is still valid b) the project is considered feature-complete with the 3.0 release and c) the software will always be available as `elephas` package. Thanks for this question, though!

@maxpumperla I think it would be better if the repo url points to the new repo, since new updated will be pushed there. What do you think?
