### 2022-10-27 #4073 @danielenricocahall
Hello @maxpumperla , sorry for the delay! Just added now :) Did I add myself correctly? https://github.com/maxpumperla/elephas/blob/mp_joss_paper/paper.md
