+++ tests/integration/test_end_to_end.py
@@ -1,4 +1,5 @@
 import random
+from functools import wraps
 from math import isclose
 
 from tensorflow.keras.optimizers import SGD
@@ -9,8 +10,34 @@ from elephas.utils.rdd_utils import to_simple_rdd
 import pytest
 import numpy as np
 
-# enumerate possible combinations for training mode and parameter server for a classification model while also validatiing
-# multiple workers for repartitioning
+
+def _generate_random_port_number():
+    return 4000 + random.randint(0, 900)
+
+
+def retry_test(stop_max_attempt_number: int):
+    def decorator(test_func_ref):
+        @wraps(test_func_ref)
+        def wrapper(*args, **kwargs):
+            retry_count = 1
+
+            while retry_count < stop_max_attempt_number:
+                try:
+                    return test_func_ref(*args, **kwargs)
+
+                except ConnectionError as connect_error:
+                    connect_message, _ = connect_error.__str__().split("\n")
+                    print(f"Retry error: \"{test_func_ref.__name__}\" --> {connect_message}. "
+                          f"[{retry_count}/{stop_max_attempt_number - 1}] Retrying.")
+                    retry_count += 1
+
+            return test_func_ref(*args, **kwargs)
+
+        return wrapper
+    return decorator
+
+# enumerate possible combinations for training mode and parameter server for a classification model while also
+# validatiing multiple workers for repartitioning
 @pytest.mark.parametrize('mode,parameter_server_mode,num_workers',
                          [('synchronous', None, None),
                           ('synchronous', None, 2),
@@ -22,6 +49,7 @@ import numpy as np
                           ('hogwild', 'http', 2),
                           ('hogwild', 'socket', None),
                           ('hogwild', 'socket', 2)])
+@retry_test(3)
 def test_training_classification(spark_context, mode, parameter_server_mode, num_workers, mnist_data, classification_model):
     # Define basic parameters
     batch_size = 64
@@ -40,7 +68,7 @@ def test_training_classification(spark_context, mode, parameter_server_mode, num
 
     # Initialize SparkModel from keras model and Spark context
     spark_model = SparkModel(classification_model, frequency='epoch', num_workers=num_workers,
-                             mode=mode, parameter_server_mode=parameter_server_mode, port=4000 + random.randint(0, 800))
+                             mode=mode, parameter_server_mode=parameter_server_mode, port=_generate_random_port_number())
 
     # Train Spark model
     spark_model.fit(rdd, epochs=epochs, batch_size=batch_size,
@@ -76,6 +104,7 @@ def test_training_classification(spark_context, mode, parameter_server_mode, num
                           ('hogwild', 'http', 2),
                           ('hogwild', 'socket', None),
                           ('hogwild', 'socket', 2)])
+@retry_test(3)
 def test_training_regression(spark_context, mode, parameter_server_mode, num_workers, boston_housing_dataset,
                              regression_model):
     x_train, y_train, x_test, y_test = boston_housing_dataset
@@ -87,7 +116,7 @@ def test_training_regression(spark_context, mode, parameter_server_mode, num_wor
     sgd = SGD(lr=0.0000001)
     regression_model.compile(sgd, 'mse', ['mae', 'mean_absolute_percentage_error'])
     spark_model = SparkModel(regression_model, frequency='epoch', mode=mode, num_workers=num_workers,
-                             parameter_server_mode=parameter_server_mode, port=4000 + random.randint(0, 800))
+                             parameter_server_mode=parameter_server_mode, port=_generate_random_port_number())
 
     # Train Spark model
[…]
