### 2022-11-08 #4073 @sepandhaghighi
@diehlpk @maxpumperla @danielenricocahall

**Result:**

A well-known and interesting tool from a prolific team 💯
But there are some concerns about **documentation** and **community guidelines**:

1. Documentation
    - I didn't find any API level documentation ([https://danielenricocahall.github.io/elephas/](https://danielenricocahall.github.io/elephas/) is available but it only contains some examples, is it sufficient? I'm not sure!!)
    - This [link](https://maxpumperla.com/elephas/) is broken
    - Some functions don't have docstring

2. Community Guidelines

    - No specific contribution guideline (without such a guideline, it's very hard to contribute to this project)
    - No issue template
    - No PR template

* @danielenricocahall is one of the paper authors, so it would be better to remove his name from the `Acknowledgments` section
* I suggest reporting your tests coverage (by something like [codecov](https://about.codecov.io/))

Best Regards
SH
