### 2022-11-14 #4073 @diehlpk
> @diehlpk happy to address the above comments, but given that we've submitted this ~10 months ago and not much happened in the meantime, I'd prefer to get a preliminary "go" from you first and only then put in more work.

I think you should address the comments. The software and paper is good and I think it deserves to be published. However, making the software user-friendly will take some effort.
