### 2022-11-14 #4073 @maxpumperla
@diehlpk sure, like I said, we definitely intend to! After maintaining this for around 6 years now, I just lean towards being a little conservative when it comes to requests.

@danielenricocahall how does your November look like? Maybe we can discuss in private how to tackle these comments? I think this all very reasonable and should help the project a lot.
