### 2022-11-15 f7b44e4
adding typehints and docstrings
