+++ .github/ISSUE_TEMPLATE/elephas-bug-report.md
@@ -1,0 +1,24 @@
+---
+name: Elephas Bug Report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: danielenricocahall
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Desktop (please complete the following information):**
+ - OS: [e.g. iOS]
+ - Version [e.g. 22]
+
+**Additional context**
+Add any other context about the problem here.
