+++ .github/ISSUE_TEMPLATE/elephas-feature-request.md
@@ -1,0 +1,20 @@
+---
+name: Elephas Feature Request
+about: Suggest an idea for this project
+title: ''
+labels: enhancement
+assignees: danielenricocahall
+
+---
+
+**Overview**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Possible Approaches**
+A clear and concise description of what you want to happen.
+
+**Alternative Options**
+A clear and concise description of any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
