### 2022-11-28 #4073 @danielenricocahall
- Added some contribution guidelines here: https://github.com/danielenricocahall/elephas/blob/master/CONTRIBUTING.md
 - Added three issue templates: https://github.com/danielenricocahall/elephas/tree/master/.github/ISSUE_TEMPLATE
- Updated some function docstrings
Will work on more in the todo list soon! I agree with the mid-December assessment.
