### 2022-12-01 danielenricocahall/elephas#15 @nmoran
Parameters passed to SGD optimizer in example deprecated in latest keras versions
**Describe the bug**
After installing elephas in a fresh python environment and trying to run the `ml_mlp_classification.py` example, I got following error

```
ValueError: decay is deprecated in the new Keras optimizer, pleasecheck the docstring for valid arguments, or use the legacy optimizer, e.g., tf.keras.optimiz
ers.legacy.SGD.
```
Would be good to update the example or pin keras to an older version in the requirements.txt.

**To Reproduce**
From clean python 3.8 environment on ubuntu 20.04
```
pip install elephas
spark-submit --driver-memory 1G ./ml_mlp_classification.py
```

**Expected behavior**
Example should run.

 - OS: ubuntu 20.04
 - Version 20.04

**Additional context**
keras==2.11.0
