### 2022-12-06 af38482
fix 'UserWarning: The `lr` argument is deprecated, use `learning_rate` instead.   super(SGD, self).__init__(name, **kwargs)'
