+++ tests/test_ml_model.py
@@ -52,7 +52,7 @@ def test_spark_ml_model_classification(spark_context, classification_model, mnis
     df = to_data_frame(spark_context, x_train, y_train, categorical=True)
     test_df = to_data_frame(spark_context, x_test, y_test, categorical=True)
 
-    sgd = optimizers.SGD(lr=0.01, decay=1e-6, momentum=0.9, nesterov=True)
+    sgd = optimizers.SGD(learning_rate=0.01, decay=1e-6, momentum=0.9, nesterov=True)
     sgd_conf = optimizers.serialize(sgd)
 
     # Initialize Spark ML Estimator
@@ -128,7 +128,7 @@ def test_regression_model(spark_context, regression_model, boston_housing_datase
     df = to_data_frame(spark_context, x_train, y_train)
     test_df = to_data_frame(spark_context, x_test, y_test)
 
-    sgd = optimizers.SGD(lr=0.00001)
+    sgd = optimizers.SGD(learning_rate=0.00001)
     sgd_conf = optimizers.serialize(sgd)
     estimator = ElephasEstimator()
     estimator.set_keras_model_config(regression_model.to_json())
@@ -165,7 +165,7 @@ def test_set_cols_deprecated(spark_context, regression_model, boston_housing_dat
         test_df = test_df.withColumnRenamed('features', 'scaled_features')
         test_df = test_df.withColumnRenamed('label', 'ground_truth')
 
-        sgd = optimizers.SGD(lr=0.00001)
+        sgd = optimizers.SGD(learning_rate=0.00001)
         sgd_conf = optimizers.serialize(sgd)
         estimator = ElephasEstimator()
         estimator.set_keras_model_config(regression_model.to_json())
@@ -204,7 +204,7 @@ def test_set_cols(spark_context, regression_model, boston_housing_dataset):
     test_df = test_df.withColumnRenamed('features', 'scaled_features')
     test_df = test_df.withColumnRenamed('label', 'ground_truth')
 
-    sgd = optimizers.SGD(lr=0.00001)
+    sgd = optimizers.SGD(learning_rate=0.00001)
     sgd_conf = optimizers.serialize(sgd)
     estimator = ElephasEstimator(labelCol='ground_truth', outputCol='output', featuresCol='scaled_features')
     estimator.set_keras_model_config(regression_model.to_json())
@@ -240,7 +240,7 @@ def test_custom_objects(spark_context, boston_housing_dataset):
     df = to_data_frame(spark_context, x_train, y_train)
     test_df = to_data_frame(spark_context, x_test, y_test)
 
-    sgd = optimizers.SGD(lr=0.00001)
+    sgd = optimizers.SGD(learning_rate=0.00001)
     sgd_conf = optimizers.serialize(sgd)
     estimator = ElephasEstimator()
     estimator.set_keras_model_config(model.to_json())
@@ -270,7 +270,7 @@ def test_predict_classes_probability(spark_context, classification_model, mnist_
     df = to_data_frame(spark_context, x_train, y_train, categorical=True)
     test_df = to_data_frame(spark_context, x_test, y_test, categorical=True)
 
-    sgd = optimizers.SGD(lr=0.01, decay=1e-6, momentum=0.9, nesterov=True)
+    sgd = optimizers.SGD(learning_rate=0.01, decay=1e-6, momentum=0.9, nesterov=True)
     sgd_conf = optimizers.serialize(sgd)
 
     # Initialize Spark ML Estimator
@@ -307,7 +307,7 @@ def test_batch_predict_classes_probability(spark_context, classification_model, 
[…]
