+++ examples/mnist_mlp_spark.py
@@ -44,7 +44,7 @@ model.add(Dropout(0.2))
 model.add(Dense(10))
 model.add(Activation('softmax'))
 
-sgd = SGD(lr=0.1)
+sgd = SGD(learning_rate=0.1)
 model.compile(sgd, 'categorical_crossentropy', ['acc'])
 
 # Build RDD from numpy features and labels
