### 2022-12-06 2c8d7a2
pin reqs to <=2.10
