### 2022-12-06 7fd13de
fix creation of sql context objects
