### 2022-12-06 649f032
remove references to sql context as they have been deprecated
