+++ elephas/ml/adapter.py
@@ -1,4 +1,3 @@
-from pyspark.sql import SQLContext
 from pyspark.mllib.regression import LabeledPoint
 from ..utils.rdd_utils import from_labeled_point, to_labeled_point, lp_to_simple_rdd
 from pyspark.mllib.linalg import Vector as MLLibVector, Vectors as MLLibVectors
@@ -7,9 +6,9 @@ from pyspark.mllib.linalg import Vector as MLLibVector, Vectors as MLLibVectors
 def to_data_frame(sc, features, labels, categorical=False):
     """Convert numpy arrays of features and labels into Spark DataFrame
     """
+    from pyspark.sql import SparkSession
     lp_rdd = to_labeled_point(sc, features, labels, categorical)
-    sql_context = SQLContext._get_or_create(sc)
-    df = sql_context.createDataFrame(lp_rdd)
+    df = SparkSession.builder.getOrCreate().createDataFrame(lp_rdd)
     return df
 
 
@@ -24,9 +23,10 @@ def from_data_frame(df, categorical=False, nb_classes=None):
 def df_to_simple_rdd(df, categorical=False, nb_classes=None, features_col='features', label_col='label'):
     """Convert DataFrame into RDD of pairs
     """
-    sql_context = df.sql_ctx
-    sql_context.registerDataFrameAsTable(df, "temp_table")
-    selected_df = sql_context.sql(
+    from pyspark.sql import SparkSession
+    spark_session = SparkSession.builder.getOrCreate()
+    df.createOrReplaceTempView("temp_table")
+    selected_df = spark_session.sql(
         "SELECT {0} AS features, {1} as label from temp_table".format(features_col, label_col))
     if isinstance(selected_df.first().features, MLLibVector):
         lp_rdd = selected_df.rdd.map(
