### 2022-12-06 danielenricocahall/elephas#16 @danielenricocahall
Hello @nmoran , thank you for the test! I have debugged the issue - after some testing, it looks like we are not compatible with the latest version of Tensorflow (2.11). As such, I have pinned the requirements to <= 2.10 at the moment. It should now be working - please test it out when you get the chance!
