### 2022-12-06 danielenricocahall/elephas#15 @danielenricocahall
Hello @nmoran , thank you for the test! Similar to the other issue, it looks like we are not compatible with Tensorflow 2.11. I have pinned the version to <=2.10 in the dependencies. Please test the latest (3.4.2) when you get the chance!
