### 2022-12-07 711f04c
fix setup.py install reqs
