+++ setup.py
@@ -2,13 +2,17 @@ from setuptools import setup
 from setuptools import find_packages
 
 setup(name='elephas',
-      version='3.4.2',
+      version='3.4.3',
       description='Deep learning on Spark with Keras',
       url='http://github.com/danielenricocahall/elephas',
       download_url='https://github.com/danielenricocahall/elephas/tarball/3.4.0',
       author='Daniel Cahall',
       author_email='danielenricocahall@gmail.com',
-      install_requires=[line.strip() for line in open('requirements.txt').readlines()],
+      iinstall_requires=['cython',
+                        'tensorflow>=2,!=2.2.*,<=2.10',
+                        'flask',
+                        'h5py==3.3.0',
+                        'pyspark<3.4'],
       extras_require={
         'tests': ['pytest', 'pytest-pep8', 'pytest-cov', 'pytest-spark', 'mock']
     },
