### 2022-12-07 danielenricocahall/elephas#16 @danielenricocahall
Correction: the latest is now 3.4.3 - 3.4.2 had an installation error due to a change in `setup.py` which I've resolved.
