### 2022-12-07 d48e7c0
fix packaging and add pypi docs
