+++ setup.py
@@ -1,14 +1,18 @@
 from setuptools import setup
 from setuptools import find_packages
-
+from pathlib import Path
+this_directory = Path(__file__).parent
+long_description = (this_directory / "README.md").read_text()
 setup(name='elephas',
-      version='3.4.3',
+      version='3.4.6',
       description='Deep learning on Spark with Keras',
-      url='http://github.com/danielenricocahall/elephas',
+      long_description=long_description,
+      long_description_content_type='text/markdown',
+    url='http://github.com/danielenricocahall/elephas',
       download_url='https://github.com/danielenricocahall/elephas/tarball/3.4.0',
       author='Daniel Cahall',
       author_email='danielenricocahall@gmail.com',
-      iinstall_requires=['cython',
+      install_requires=['cython',
                         'tensorflow>=2,!=2.2.*,<=2.10',
                         'flask',
                         'h5py==3.3.0',
