### 2022-12-07 #4073 @danielenricocahall
I believe I addressed the issues in the latest release (3.4.6), which required pinning the Tensorflow version to <=2.10. Also, PyPI now has the Github documentation: https://pypi.org/project/elephas/3.4.6/.
