### 2022-12-10 #4073 @nmoran
The paper and software look good to me and is of use to many users as evidenced by the large number of monthly downloads 💯

I am happy to recommend this be published in JOSS.

In addition I think it could be useful to mention the versions of python that elephas was developed and tested with to save users time in case there are unforeseen incompatibilities with older or newer versions.
