### 2022-12-13 #4073 @maxpumperla
```
affiliations:
  - name: IU Internationale Hochschule, Erfurt, Germany
    index: 1
  - name: Anyscale Inc, San Francisco, USA
    index: 2
```

@danielenricocahall I don't know which branch and repo we settled on for the PDF by now, but could you please update affiliation info for me like this when you update yours? thank you.

also, many thanks @diehlpk for finally moving this towards the finish line!!!
