### 2022-12-14 #4073 @diehlpk
@danielenricocahall the title of the archive is wrong and it should read

Elephas: Distributed Deep Learning with Keras & Spark

You can edit the title and there is no need for a new DOI.
