### 2022-12-15 #4073 @danielenricocahall
Also, just realized we didn't change the API documentation link to https://danielenricocahall.github.io/elephas/ instead of the old page (http://maxpumperla.com/elephas/). Is there any way we could edit that too?
