### 2022-12-15 #4073 @arfon
> Wooo! This is exciting! Everything looks good except on the DOI it looks like it links to the old repository - is there any way to edit that? I'm not sure where that was set or if we missed something.

Could you be be more specific here, i.e., what link is wrong and where?

> Also, just realized we didn't change the API documentation link to https://danielenricocahall.github.io/elephas/ instead of the old page (http://maxpumperla.com/elephas/). Is there any way we could edit that too?

Yes, just update your paper and we can re-accept
