### 2022-12-15 #4073 @danielenricocahall
Sure! Specifically, I mean on https://joss.theoj.org/papers/10.21105/joss.04073, the "Software Repository" button links to https://github.com/maxpumperla/elephas instead of https://github.com/danielenricocahall/elephas, the currently active fork. It's linked correctly in the actual paper (the "Repository" link goes to https://github.com/danielenricocahall/elephas). And sure thing, will edit now.
