### 2022-12-15 #4073 @danielenricocahall
> @danielenricocahall – can you take a look and see if things are looking better for the doc now?

@arfon checking now. The link is correct now (thank you!). It looks like the hyperlinks are still out-of-date though. Is it because the new pdf wasn't successfully published before the reaccept?
